refactor: separate matching service logic into files
