+++ backend/matching-service/app/.env.example
@@ -1,4 +1,4 @@
 NODE_ENV=development
-PORT=8004
+PORT=80
 FRONTEND_ORIGIN=http://localhost:8000
 API_GATEWAY_URL=http://localhost:8001
