+++ backend/matching-service/app/src/matching/matching.constants.ts
@@ -1,4 +1,4 @@
 export const MATCHING_EVENTS = {
   JOIN_ROOM: "join_room",
   JOINED_ROOM: "joined_room",
-} as const;
+};
