+++ frontend/app/.env.development
@@ -1,1 +1,2 @@
 VITE_BACKEND_URL=http://localhost:8001
+VITE_MATCHING_SERVICE_URL=http://localhost:8004
