+++ frontend/app/src/lib/env.ts
@@ -2,7 +2,7 @@ import zod from "zod";
 
 const envSchema = zod.object({
   VITE_BACKEND_URL: zod.string().url().nonempty(),
-  MATCHING_SERVICE_URL: zod.string().url().nonempty(),
+  VITE_MATCHING_SERVICE_URL: zod.string().url().nonempty(),
 });
 
 export const env = envSchema.parse(import.meta.env);
